03 End To End Resume Application Tracking System(ATS) Using Google Gemini Pro Visio/requirements.txt
# ---- UI Framework ----
streamlit              # Simple web app frontend to upload resumes and display results

# ---- Google Gemini API ----
google-generativeai     # Official Gemini Pro SDK for text and vision models (Pro Vision)

# ---- Environment Variables ----
python-dotenv           # Load your API keys (like Google API) securely from .env file

# ---- PDF to Image Conversion ----
pdf2image              # Convert PDF resume pages to images for Gemini Vision input


